requirements.txt
streamlit==1.45.1
pandas==2.3.0
python-dotenv==1.1.0
groq==0.31.0
semantic-router==0.0.27
chromadb==1.0.16
sentence_transformers==5.1.0
cohere==4.57